+++ utils/slide/src/utils/component-utils.ts
