+++ utils/slide/src/utils/images-utils.ts
